+++ example/lib/main.dart
@@ -1,9 +1,8 @@
 import 'package:flutter/material.dart';
 import 'dart:async';
-
+import 'package:nepali_calendar/nepali_date_converter.dart';
+import 'package:nepali_calendar/nepali_date_picker.dart';
 import 'package:flutter/services.dart';
-import 'package:nepali_calendar/nepali_calendar.dart';
-import 'package:nepali_calendar_example/nepali_date_converter.dart';
 
 void main() => runApp(MyApp());
 
@@ -13,26 +12,43 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> {
+  @override
+  void initState() {
+    super.initState();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      home: MainScreen(),
+    );
+  }
+}
+
+class MainScreen extends StatefulWidget {
+  @override
+  _MainScreenState createState() => _MainScreenState();
+}
+
+class _MainScreenState extends State<MainScreen> {
   String nepaliDate = "";
   String convertedEnglish = "";
   String englishDate = "";
   String convertedNepali = "";
 
   @override
-  void initState() {
-    super.initState();
-  }
-  
-  @override
   Widget build(BuildContext context) {
-    return MaterialApp(
-      home: Scaffold(
-        appBar: AppBar(
-          title: const Text('Plugin example app'),
-        ),
-        body: Column(
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('Plugin example app'),
+      ),
+      body: Container(
+        margin: EdgeInsets.all(16),
+        child: ListView(
           children: <Widget>[
             TextField(
+              keyboardType: TextInputType.number,
+              inputFormatters: [DateTextInputFormatter()],
               decoration: InputDecoration(hintText: "BS - (yyyy-MM-dd)"),
               onChanged: (value) => this.nepaliDate = value,
             ),
@@ -65,6 +81,8 @@ class _MyAppState extends State<MyApp> {
             ),
             TextField(
                 decoration: InputDecoration(hintText: "AD - (yyyy-MM-dd)"),
+                inputFormatters: [DateTextInputFormatter()],
+                keyboardType: TextInputType.number,
                 onChanged: (value) => this.englishDate = value),
             Text(convertedNepali),
             RaisedButton(
@@ -93,15 +111,56 @@ class _MyAppState extends State<MyApp> {
                 }
               },
             ),
+            SizedBox(
+              height: 20,
+            ),
+            Text(pickedDate != null
+                ? "${pickedDate.toString()} => ${pickedDate.dateTime.year}-${pickedDate.dateTime.month}-${pickedDate.dateTime.day}"
+                : "Pick a date"),
+            RaisedButton(
+              child: Text("PickDate"),
+              onPressed: () {
+                NepaliDatePicker(context, currentDate: pickedDate,
+                    onSelected: (nepalidate) {
+                  print(nepalidate.toString());
+                  print(nepalidate.dateTime);
+                  setState(() {
+                    pickedDate = nepalidate;
+                  });
+                }).showDatePicker();
+              },
+            )
           ],
         ),
-        floatingActionButton: FloatingActionButton(
-          onPressed: () {
-            NepaliDate.fromBS(2076, 1, 5);
-            NepaliDate.fromAD(DateTime(2019, 4, 14));
-          },
-        ),
       ),
+    );
+  }
+
+  NepaliDate pickedDate;
+}
+
+class DateTextInputFormatter extends TextInputFormatter {
+  @override
+  TextEditingValue formatEditUpdate(
+      TextEditingValue oldValue, TextEditingValue newValue) {
+    final int newTextLength = newValue.text.trim().length;
+    String newText;
+    if (newTextLength < oldValue.text.trim().length) {
+      newText = newValue.text.endsWith("-")
+          ? newValue.text.substring(0, newValue.text.length - 1)
+          : newValue.text;
+    } else if (newTextLength > 10) {
+      newText = oldValue.text;
+    } else if (newTextLength == 5 || newTextLength == 8) {
+      newText = newValue.text.substring(0, newTextLength - 1) +
+          "-" +
+          newValue.text.substring(newTextLength - 1);
+    } else
+      newText = newValue.text;
+    newText = newText.replaceAll("--", "-");
+    return TextEditingValue(
+      text: newText,
+      selection: TextSelection.collapsed(offset: newText.length),
     );
   }
 }